convert last Java classes in Kotlin

+++ src/main/java/com/fouan/events/GameStarted.kt
@@ -1,3 +1,3 @@
 package com.fouan.events
 
-class GameStarted(turn: Int, private val timestamp: Int) : Event<GameStarted>(turn)
+class GameStarted(turn: Int, private val timestamp: Long) : Event<GameStarted>(turn)
